refatoração: reorganiza estrutura da camada de apresentação
- Remove pasta 'pages' e unifica todas as telas em 'presentation/screens'
- Mantém componentes reutilizáveis em 'presentation/widgets'
- Atualiza imports para refletir a nova organização
- Melhora a organização e facilita a manutenção do projeto
- Prepara o projeto para maior escalabilidade e facilidade de onboarding
